Refactor translation scripts to share common logic via translation_utils.py
