+++ scripts/check_translation_values.py
@@ -4,61 +4,208 @@ Script to check for potential translation leaks (untranslated values) in languag
 This complements check_trn_keys.py by verifying the *content* of translations, not just the keys.
 """
 
-import json
-import glob
 import os
 import sys
 
-# Configuration
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-LANG_DIR = os.path.join(PROJECT_ROOT, "src", "assets", "lang")
+from translation_utils import get_json_files, load_json
 
-def load_json(path):
-    with open(path, 'r', encoding='utf-8') as f:
-        return json.load(f)
 
 def main():
     print("=== Translation Content Check (Value Verification) ===\n")
 
-    json_files = glob.glob(os.path.join(LANG_DIR, "*.json"))
+    json_files = get_json_files()
     has_error = False
 
     # Technical terms or common untranslated strings to ignore in warnings
     # These are valid to be identical in many languages (e.g. units, acronyms)
     IGNORED_IDENTICALS = {
         # Units
-        "Hz", "kHz", "MHz", "dB", "dBFS", "dBV", "dBu", "V", "mV", "FS",
-        "s", "ms", "μs", "ns", "%", "ppm", "deg", "Ω",
-
+        "Hz",
+        "kHz",
+        "MHz",
+        "dB",
+        "dBFS",
+        "dBV",
+        "dBu",
+        "V",
+        "mV",
+        "FS",
+        "s",
+        "ms",
+        "μs",
+        "ns",
+        "%",
+        "ppm",
+        "deg",
+        "Ω",
         # Modulation / Signals
-        "AM", "FM", "PM", "ΦM", "DSB", "SSB", "LSB", "USB", "LO", "CW",
-        "SINAD", "THD", "THD+N", "IMD", "SNR", "ENOB", "SFDR",
-        "RMS", "Pk", "pp", "CF", "Lp", "Leq", "SEL", "SPL", "dB SPL",
-        "LUFS", "LUFS (M)", "LUFS (S)", "LUFS (I)",
-        "SMPTE", "CCIF", "DIN", "ISO", "IEC", "ITU-R", "EBU",
-        "1/f", "White", "Pink", "Brown",
-
+        "AM",
+        "FM",
+        "PM",
+        "ΦM",
+        "DSB",
+        "SSB",
+        "LSB",
+        "USB",
+        "LO",
+        "CW",
+        "SINAD",
+        "THD",
+        "THD+N",
+        "IMD",
+        "SNR",
+        "ENOB",
+        "SFDR",
+        "RMS",
+        "Pk",
+        "pp",
+        "CF",
+        "Lp",
+        "Leq",
+        "SEL",
+        "SPL",
+        "dB SPL",
+        "LUFS",
+        "LUFS (M)",
+        "LUFS (S)",
+        "LUFS (I)",
+        "SMPTE",
+        "CCIF",
+        "DIN",
+        "ISO",
+        "IEC",
+        "ITU-R",
+        "EBU",
+        "1/f",
+        "White",
+        "Pink",
+        "Brown",
         # Hardware / Software
-        "CPU", "RAM", "GPU", "OS", "GUI", "API", "ASIO", "WASAPI", "ALSA", "JACK",
-        "PCI", "LAN", "IP", "TCP", "UDP",
-        "1 PPS", "10 MHz", "NCO", "DDS", "DAC", "ADC", "FPGA", "DSP",
-        "FIR", "IIR", "SOS", "FFT", "DFT", "PDF", "CDF",
-        "L", "R", "M", "S", "X", "Y", "Z", "I", "Q",
-        "CH1", "CH2", "Ch 1", "Ch 2", "Ch 1 (L)", "Ch 2 (R)",
-        "Stereo", "Mono", "Left", "Right", "Mid", "Side",
-        "Input", "Output", "I/O", "Ref", "Meas", "Trig", "Gate",
-
+        "CPU",
+        "RAM",
+        "GPU",
+        "OS",
+        "GUI",
+        "API",
+        "ASIO",
+        "WASAPI",
+        "ALSA",
+        "JACK",
+        "PCI",
+        "LAN",
+        "IP",
+        "TCP",
+        "UDP",
+        "1 PPS",
+        "10 MHz",
+        "NCO",
+        "DDS",
+        "DAC",
+        "ADC",
+        "FPGA",
+        "DSP",
+        "FIR",
+        "IIR",
+        "SOS",
+        "FFT",
+        "DFT",
+        "PDF",
+        "CDF",
+        "L",
+        "R",
+        "M",
+        "S",
+        "X",
+        "Y",
+        "Z",
+        "I",
+        "Q",
+        "CH1",
+        "CH2",
+        "Ch 1",
+        "Ch 2",
+        "Ch 1 (L)",
+        "Ch 2 (R)",
+        "Stereo",
+        "Mono",
+        "Left",
+        "Right",
+        "Mid",
+        "Side",
+        "Input",
+        "Output",
+        "I/O",
+        "Ref",
+        "Meas",
+        "Trig",
+        "Gate",
         # UI Common
-        "OK", "Cancel", "Apply", "Reset", "Save", "Load", "Import", "Export",
-        "Start", "Stop", "Pause", "Resume", "Run", "Ready", "Idle", "Busy",
-        "Error", "Warning", "Info", "System", "Settings", "Help", "About",
-        "Name", "Value", "Type", "Mode", "Status", "Time", "Date",
-        "Min", "Max", "Avg", "Mean", "Std Dev", "Count", "Rate", "Speed",
-        "Gain", "Phase", "Freq", "Level", "Offset", "Delay", "Width", "Depth",
-        "Buffer", "Sample Rate", "Bit Depth", "Channels",
-        "General", "Audio", "Appearance", "Language", "Theme",
-        "Dark", "Light", "Auto", "Manual", "Custom", "Default",
-        "N/A", "-", "--", "---", "..."
+        "OK",
+        "Cancel",
+        "Apply",
+        "Reset",
+        "Save",
+        "Load",
+        "Import",
+        "Export",
+        "Start",
+        "Stop",
+        "Pause",
+        "Resume",
+        "Run",
+        "Ready",
+        "Idle",
+        "Busy",
+        "Error",
+        "Warning",
+        "Info",
+        "System",
+        "Settings",
+        "Help",
+        "About",
+        "Name",
+        "Value",
+        "Type",
+        "Mode",
+        "Status",
+        "Time",
+        "Date",
+        "Min",
+        "Max",
+        "Avg",
+        "Mean",
+        "Std Dev",
+        "Count",
+        "Rate",
+        "Speed",
+        "Gain",
+        "Phase",
+        "Freq",
+        "Level",
+        "Offset",
+        "Delay",
+        "Width",
+        "Depth",
+        "Buffer",
+        "Sample Rate",
+        "Bit Depth",
+        "Channels",
+        "General",
+        "Audio",
+        "Appearance",
+        "Language",
+        "Theme",
+        "Dark",
+        "Light",
+        "Auto",
+        "Manual",
+        "Custom",
+        "Default",
+        "N/A",
+        "-",
+        "--",
+        "---",
+        "...",
     }
 
     # Dynamic heuristics for valid identicals
@@ -68,7 +215,7 @@ def main():
 
     for jf in sorted(json_files):
         fname = os.path.basename(jf)
-        if fname == 'en.json':
+        if fname == "en.json":
             continue
 
         try:
@@ -115,15 +262,15 @@ def main():
         if empty_values:
             print(f"FAIL: {fname} has {len(empty_values)} EMPTY values (Missing Translation):")
             for k in empty_values:
-                print(f"  - \"{k}\"")
+                print(f'  - "{k}"')
             has_error = True
 
         if identical_values:
             print(f"WARNING: {fname} has {len(identical_values)} potentially untranslated values (Key == Value):")
             for k in identical_values[:10]:
-                print(f"  - \"{k}\"")
+                print(f'  - "{k}"')
             if len(identical_values) > 10:
-                print(f"  ... and {len(identical_values)-10} more.")
+                print(f"  ... and {len(identical_values) - 10} more.")
 
         if not empty_values and not identical_values:
             print(f"OK: {fname}")
@@ -138,6 +285,7 @@ def main():
     else:
         print("TEST PASSED: No critical issues found.")
         sys.exit(0)
+
 
 if __name__ == "__main__":
     main()
